refactor: rename requireLogin middleware
